자동 매수, 매도 기능으로써 처리
Model <-> Protocol 변환 ModelsExtend로 이전

+++ Server/Protocols/Request/MockInvestAnalysisBuy.cs
@@ -16,6 +16,6 @@ namespace Server.Protocols.Request
 
         public int Count { get; set; }
 
-        public int TotalPrice { get; set; }
+        public long TotalPrice { get; set; }
     }
 }
